% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect}
\alias{collect}
\alias{collect.BasicRaster}
\title{Raster methods for collect}
\usage{
collect(x, ...)

\method{collect}{BasicRaster}(x, ..., nrows = 512, ncols = nrows,
  uselazy = TRUE)
}
\arguments{
\item{x}{raster object}

\item{...}{ignored}

\item{nrows}{number of rows in the output}

\item{ncols}{number of columns in the output}

\item{uselazy}{use lazy read via vapour package}
}
\description{
Raster methods for collect
}
