% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\name{crop}
\alias{crop}
\title{Lazy crop}
\description{
Set an active window of data using the same \code{\link[=crop]{crop()}} function as the
raster package.
}
\details{
This is the data window that will be pulled by conversion
to an actual raster by \code{\link[=as_raster]{as_raster()}}.
}
\examples{
sstfile <- system.file("extdata/sst.tif", package = "vapour")
lr <- lazyraster(sstfile)
## crop and stay as lazyraster
crop(lazyraster(sstfile), raster::extent(142, 143, -50, -45))
}
\seealso{
\code{\link[=lazyraster]{lazyraster()}}, and \code{\link[raster:crop]{raster::crop()}}
}
