% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyraster.R
\name{lazycrop}
\alias{lazycrop}
\alias{lazycrop.lazyraster}
\title{Lazy crop}
\usage{
lazycrop(x, y, ..., verbose = FALSE)

\method{lazycrop}{lazyraster}(x, y, ..., verbose = FALSE)
}
\arguments{
\item{x}{Raster* object or SpatialPolygons*, SpatialLines*, or SpatialPoints* object}

\item{y}{Extent object, or any object from which an Extent object can be extracted (see Details)}

\item{...}{Additional arguments as for \code{\link{writeRaster}}}
}
\description{
Set an active window of data that will be pulled by conversion
to an actual raster.
}
