% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy-methods.R
\name{lazyraster-plot-method}
\alias{lazyraster-plot-method}
\alias{plot.lazyraster}
\alias{plot}
\title{Lazy raster S3 plot method}
\usage{
plot(x, y, ...)
}
\arguments{
\item{x}{a `lazyraster`}

\item{y}{ignored}

\item{...}{passed to raster plot}
}
\description{
Plot for lazyraster.
}
