% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00lazy-S4-methods.R
\name{lazyraster-raster-S4}
\alias{lazyraster-raster-S4}
\alias{extent}
\alias{crop}
\alias{raster_lazyraster}
\title{Raster methods (S4) for lazyraster.}
\usage{
raster_lazyraster(x, ...)
}
\arguments{
\item{x}{various (raster, extent)}

\item{...}{arguments passed to underlying raster function}
}
\description{
These methods are generics from the raster package, extended to work for lazyrasters.
}
\seealso{
\code{\link[raster:raster]{raster::raster()}}, \code{\link[raster:extent]{raster::extent()}}
}
