% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazywarp.R
\name{lazywarp}
\alias{lazywarp}
\title{Simple raster warping in-memory}
\usage{
lazywarp(gdalsource, target, band = 1L, a_srs = NULL)
}
\arguments{
\item{gdalsource}{a file name or other source string}

\item{target}{raster target object}

\item{band}{band number (single) of source}

\item{a_srs}{PROJ4 string to apply to gdalsource in case it is wrong or missing}
}
\description{
Warp a GDAL data source to in in-memory raster target
}
\details{
Currently only works as if 64-bit real band values, but GDAL's RasterIO
does sort out conversions. We set a missing value as if it's very low, but
it might be zero (as in the example), so WIP.
}
\examples{
laea_srs <- "+proj=laea +lon_0=147 +lat_0=-52 +datum=WGS84"
ex <- c(-1e6, 1e6, -1e6, 1e6)
library(raster)
ras <- raster(extent(ex), crs = laea_srs, nrows= 100, ncols = 100)
f <- system.file("extdata/sst.tif", package = "vapour", mustWork = TRUE)
#r <- projectRaster(raster(f), ras)
r2 <- lazywarp(f, ras)
r2[r2 < 1] <- NA
plot(r2)
}
