% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merc.R
\name{merc}
\alias{merc}
\title{Mercator extent}
\usage{
merc(x = 146.7, y = -42, wh = 256000)
}
\arguments{
\item{x}{longitude}

\item{y}{latitude}

\item{wh}{width and height, in metres - if only one vaue provided it is also used for height}
}
\value{
four values, xmin, xmax, ymin, ymax in global Mercator
}
\description{
Create an extent in Mercator projection from a longitude, latitude, and a width,height.
}
\examples{
merc()  ## a parochial default
library(lazyraster)
vearth <- '<GDAL_WMS> <Service name="VirtualEarth">
<ServerUrl>http://a${server_num}.ortho.tiles.virtualearth.net/tiles/a${quadkey}.jpeg?g=90</ServerUrl></Service>
<MaxConnections>4</MaxConnections>    <Cache/>    </GDAL_WMS>'
lr <- lazyraster(vearth)
\dontrun{\donttest{
raster::plotRGB(as_raster(crop(lr, merc(-90, 52, 256e4)), band = 1:3))
}}
}
