% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect}
\alias{collect}
\alias{collect}
\alias{collect.BasicRaster}
\title{Raster methods for collect}
\usage{
collect(x, ...)

\method{collect}{BasicRaster}(x, ..., nrows = 512, ncols = nrows)
}
\description{
Raster methods for collect
}
