% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyraster.R
\name{lazyraster-methods}
\alias{lazyraster-methods}
\alias{print.lazyraster}
\alias{lazyraster-methods}
\alias{extent.lazyraster}
\alias{lazyraster-methods}
\alias{format.lazyraster}
\alias{lazyraster-methods}
\alias{print.lazyraster}
\alias{lazyraster-methods}
\alias{plot.lazyraster}
\alias{lazyraster-methods}
\alias{raster.lazyraster}
\title{Lazy raster S3 methods}
\usage{
\method{print}{lazyraster}(x, ...)

extent.lazyraster(x, ...)

\method{format}{lazyraster}(object, ...)

\method{print}{lazyraster}(x, ...)

plot.lazyraster(x, ...)

raster.lazyraster(x, ...)
}
\arguments{
\item{x}{a `lazyraster`}

\item{...}{reserved}
}
\description{
Lazy raster S3 methods
}
