% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy-methods.R
\name{lazyraster-methods}
\alias{lazyraster-methods}
\alias{print.lazyraster}
\alias{lazyraster-methods}
\alias{format.lazyraster}
\title{Lazy raster S3 methods}
\usage{
\method{print}{lazyraster}(x, ...)

\method{format}{lazyraster}(x, ...)
}
\arguments{
\item{x}{a `lazyraster`}

\item{...}{reserved}
}
\description{
Print and format for lazyraster.
}
