% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyraster.R
\name{lazyraster}
\alias{lazyraster}
\alias{lazyraster}
\alias{as_raster}
\alias{lazyraster}
\alias{as_raster}
\alias{lazyraster}
\title{Lazy raster}
\usage{
lazyraster(gdalsource)

as_raster(x, dim = NULL, resample = "NearestNeighbour")

as_raster(x, dim = NULL, resample = "NearestNeighbour")
}
\arguments{
\item{gdalsource}{a file name or other source string}

\item{x}{a lazyraster}

\item{dim}{dimensions of data to return}

\item{resample}{resample method to use, see `vapour::raster_io`}
}
\description{
Read metadata only from a raster source, for later use with plotting and
conversion to raster.

# importFrom raster crop
# export crop
# name lazyraster-methods
crop.lazyraster <- function(x, y, ...) {
  stop("cannot crop a lazyraster, use lazycrop")
}
}
\examples{
sstfile <- system.file("extdata/sst.tif", package = "vapour")
lazyraster(sstfile)
as_raster(lazyraster(sstfile))
as_raster(lazycrop(lazyraster(sstfile), extent(142, 143, -50, -45)))
}
